third_party/blink/web_tests/fast/css/invalidation/any-link-pseudo-expected.txt

Use descendant invalidation set for :any-link pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(link).backgroundColor is red
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(link).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

This link should have a green background.
